Move SortBySelect into own file
